test : 주문 승인록직 한번만 호출 검증
